Rename commandExecutor to redocExecutor
